Add entrypoint
:man_facepalming:
